pyproject.toml
[project]
name = "instagram-personal-mcp"
version = "0.1.0"
description = "Model Context Protocol server for personal Instagram accounts (unofficial private API via instagrapi)."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Syed Aleem" },
]
keywords = ["mcp", "instagram", "instagrapi", "claude", "anthropic", "fastmcp"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "mcp[cli]>=1.2.0",
    "instagrapi>=2.0.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.6",
]

[project.urls]
Homepage = "https://github.com/AleemHaider/instagram-personal-mcp"
Repository = "https://github.com/AleemHaider/instagram-personal-mcp"
Issues = "https://github.com/AleemHaider/instagram-personal-mcp/issues"

[project.scripts]
instagram-personal-mcp = "instagram_personal_mcp.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/instagram_personal_mcp"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
